Move the description-to-handler bridge into the spine both front-ends compile
RequestModelBuilder turned an OperationModel into the RequestHandlerModel the emitters
consume, and lived beside the description reader because only described applications needed
it. Code-first is moving onto the same model, and a bridge compiled into one generator is a
bridge the other cannot cross.

The half that reads HandlerInfo stays where it was - that type describes a [Handler] class
the build task found and is internal to the Idl generator. RequestModelBuilder keeps it,
plus forwarders, so no caller or test changed.

SpecBridge/ rather than Requests/**, which was the first attempt and reached too far: every
generator wrapper links that folder, so Hardened.Function.SourceGenerator started compiling
a description model it has no descriptions for. Only the two front-ends link SpecBridge.

BuildHandlerModel now prefers OperationSymbols for the declaring and handler types. A
description names neither - both are derived by convention from a service name - and an
application that declared its own controller cannot have them guessed: the handler class is
named for a controller nobody wrote down in the model.

Nothing moved. 5,089 tests green, CI-clean under TreatWarningsAsErrors, and re-recording
every pipeline fixture changes no file.
