+++ src/SourceGenerators/Hardened.Web.SourceGenerator/Hardened.Web.SourceGenerator.csproj
@@ -56,6 +56,39 @@
         specification-first emitter compresses the same way.
     -->
     <ItemGroup>
+        <!--
+          The bridge from a description's OperationModel to a RequestHandlerModel, and the model it
+          reads. In its own folder rather than Requests/** because every generator wrapper links
+          that one - including Hardened.Function.SourceGenerator, which has no descriptions and no
+          reason to carry the model for them.
+        -->
+        <Compile Include="../Hardened.SourceGenerator/SpecBridge/**/*.cs">
+            <Link>SpecBridge/%(RecursiveDir)%(FileName)%(Extension)</Link>
+        </Compile>
+        <!--
+          The description model itself, which SpecBridge above reads.
+        -->
+        <Compile Include="../Hardened.Idl.Shared/Models/**/*.cs">
+            <Link>Idl/Models/%(RecursiveDir)%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/TypeMapper.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/NamingHelper.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/DefaultLiteral.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/SchemaShape.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/EnumWireForm.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
+        <Compile Include="../Hardened.Idl.Shared/DefaultErrorBody.cs">
+            <Link>Idl/%(FileName)%(Extension)</Link>
+        </Compile>
         <Compile Include="../Hardened.Idl.Shared/GZipLiteral.cs">
             <Link>Shared/GZipLiteral.cs</Link>
         </Compile>
